完成了登录功能的开发，实现了基于Spring-Security的登录校验，确保员工管理页面和部门管理页面必须登录后才能访问； 实现了会话跟踪技术，采用jwt令牌，这样每次请求都携带jwt令牌校验身份；

+++ src/main/java/com/gzh/tilas/Service/impl/DeptServiceImpl.java
@@ -2,8 +2,14 @@ package com.gzh.tilas.Service.impl;
 
 
 import com.gzh.tilas.Mapper.DeptMapper;
+import com.gzh.tilas.Mapper.EmployeeMapper;
 import com.gzh.tilas.Service.DeptService;
+import com.gzh.tilas.exception.BusinessException;
 import com.gzh.tilas.pojo.Dept;
+import com.gzh.tilas.pojo.DeptAddDTO;
+import com.gzh.tilas.pojo.DeptUpdateDTO;
+import com.gzh.tilas.pojo.PageBean;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +19,8 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;  // 注入Mapper接口, 用于调用Mapper接口中的方法, 访问数据库
+    @Autowired
+    private EmployeeMapper employeeMapper;
 
     // 查询所有部门信息
     @Override
@@ -27,9 +35,11 @@ public class DeptServiceImpl implements DeptService {
      * @return List<Dept>
      */
     @Override
-    public List<Dept> listByPage(Integer page, Integer pageSize) {
+    public PageBean<Dept> listByPage(Integer page, Integer pageSize) {
         int offset = (page - 1) * pageSize;
-        return deptMapper.listByPage(offset, pageSize);
+        Long total = deptMapper.count();
+        List<Dept> records = deptMapper.listByPage(offset, pageSize);
+        return new PageBean<Dept>(total, records);
     }
 
     // 根据id查询部门信息
@@ -40,19 +50,27 @@ public class DeptServiceImpl implements DeptService {
 
     // 添加部门信息
     @Override
-    public int insert(Dept dept) {
-        return deptMapper.insert(dept);
+    public int insert(DeptAddDTO deptAddDTO) {
+        Dept deptToAdd = new Dept();
+        BeanUtils.copyProperties(deptAddDTO, deptToAdd);
+        return deptMapper.insert(deptToAdd);
     }
 
-    // 修改部门信息
+    // 根据前端传过来的DeptUpdateDTO对象修改部门信息
     @Override
-    public int update(Dept dept) {
-        return deptMapper.update(dept);
+    public int update(DeptUpdateDTO deptUpdateDTO) {
+        return deptMapper.update(deptUpdateDTO);
     }
 
-    // 删除部门信息
+    // 根据部门ID删除部门信息，但是前提是该部门下已经不存在员工，否则抛出异常
     @Override
     public int deleteById(Integer id) {
+        // 检查部门下是否有员工
+        if (!employeeMapper.listByDeptId(id).isEmpty()) {
+            // 直接抛出业务异常，剩下的交给全局异常处理器
+            throw new BusinessException("该部门下尚有员工，无法删除");
+        }
+
         return deptMapper.deleteById(id);
     }
 }
